% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_core_functions}
\alias{get_core_functions}
\title{List of all packages in the SportsDataverse}
\usage{
get_core_functions()
}
\value{
A character vector of the CRAN packages that
\code{library(sportsdataverse)} attaches.
}
\description{
List of all packages in the SportsDataverse
}
\examples{
get_core_functions()
}
