% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{sportsdataverse_packages}
\alias{sportsdataverse_packages}
\title{List all packages in the sportsdataverse}
\usage{
sportsdataverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include sportsdataverse in the list?}
}
\value{
A sorted character vector of package names.
}
\description{
List all packages in the sportsdataverse
}
\examples{
sportsdataverse_packages()
}
