% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{sportsdataverse_update}
\alias{sportsdataverse_update}
\title{Update sportsdataverse packages}
\usage{
sportsdataverse_update(
  recursive = FALSE,
  repos = getOption("repos"),
  devel = FALSE
)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
sportsdataverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}

\item{devel}{If \code{TRUE}, look for prebuilt development versions on the
SportsDataverse r-universe (\url{https://sportsdataverse.r-universe.dev}) so it
is possible to install development versions without GitHub.}
}
\value{
Returns \code{NULL} invisibly. Called for its side effects.
}
\description{
This will check to see if all sportsdataverse packages (and optionally, their
dependencies) are up-to-date, and prints the install command needed to bring
any out-of-date packages current.
}
\examples{
\donttest{
try(sportsdataverse_update())
}
}
