% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_core_functions}
\alias{get_core_functions}
\title{List of all packages in the SportsDataverse}
\usage{
get_core_functions()
}
\value{
Returns a vector of the CRAN packages in the SportsDataverse
}
\description{
List of all packages in the SportsDataverse
}
