% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportsdataverse-package.R
\docType{package}
\name{sportsdataverse-package}
\alias{sportsdataverse}
\alias{sportsdataverse-package}
\title{sportsdataverse: Easily Install and Load the 'SportsDataverse'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Installs and loads, in a single step, the set of 'sportsdataverse' sports data packages, which work in harmony because they share common data representations and 'API' design. Also reports and updates the installed versions of those packages. Learn more about the 'sportsdataverse' at \url{https://sportsdataverse.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sportsdataverse/sportsdataverse-R}
  \item \url{https://r.sportsdataverse.org}
  \item Report bugs at \url{https://github.com/sportsdataverse/sportsdataverse-R/issues}
}

}
\author{
\strong{Maintainer}: Saiem Gilani \email{saiem.gilani@gmail.com} (\href{https://orcid.org/0000-0002-7194-9067}{ORCID}) [copyright holder]

}
\keyword{internal}
