% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportsdataverse-package.R
\docType{package}
\name{sportsdataverse-package}
\alias{sportsdataverse}
\alias{sportsdataverse-package}
\title{sportsdataverse: Easily Install and Load the 'SportsDataverse'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'sportsdataverse' is a set of sports data packages that work in harmony because they share common data representations and 'API' design. This package is designed to make it easy to install and load multiple 'sportsdataverse' packages in a single step. Learn more about the 'sportsdataverse' at \url{https://sportsdataverse.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saiemgilani/sportsdataverse-R}
  \item Report bugs at \url{https://github.com/saiemgilani/sportsdataverse-R/issues}
}

}
\author{
\strong{Maintainer}: Saiem Gilani \email{saiem.gilani@gmail.com}

}
\keyword{internal}
