% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{sportsdataverse_deps}
\alias{sportsdataverse_deps}
\title{List all sportsdataverse dependencies}
\usage{
sportsdataverse_deps(
  recursive = TRUE,
  pkg_list = get_core_functions(),
  repos = getOption("repos")
)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
sportsdataverse packages.}

\item{pkg_list}{The list of the package names to check (including the \code{sportsdataverse} package itself)
Defaults to \code{getOptions("repos")}.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all sportsdataverse dependencies
}
\examples{
  try(sportsdataverse_deps() \%>\% knitr::kable())

}
