% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{sportsdataverse_packages}
\alias{sportsdataverse_packages}
\title{List all packages in the nflverse}
\usage{
sportsdataverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include nflverse in the list?}
}
\description{
List all packages in the nflverse
}
\examples{
sportsdataverse_packages()
}
