% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{sportsdataverse_update}
\alias{sportsdataverse_update}
\title{Update sportsdataverse packages}
\usage{
sportsdataverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
sportsdataverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all sportsdataverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
sportsdataverse_update()
}
}
